[Quest/Malawi Core] In the app, sort the order of the tasks on the patient profile with 'double' values (#2096)

+++ android/engine/src/main/java/org/smartregister/fhircore/engine/data/local/register/dao/HivRegisterDao.kt
@@ -159,7 +159,7 @@ constructor(
           .activeTasks()
           .sortedWith(
             compareBy<Task>(
-              { it.clinicVisitOrder(metaCodingSystemTag) ?: Integer.MAX_VALUE },
+              { it.clinicVisitOrder(metaCodingSystemTag) ?: Double.MAX_VALUE },
               // tasks with no clinicVisitOrder, would be sorted with Task#description
               { it.description }
             )
